+++ lib/http/options.ts
@@ -22,7 +22,7 @@ export function createHttpOptionsConstructor() {
     headers: object;
     httpRequestClient: HttpRequestClient;
     httpRequestInterceptors: ((request: RequestOptions) => void)[];
-    enablePollDelay: boolean;
+    pollDelay: number;
     
     constructor(args: any) {
       super(args);
@@ -31,7 +31,7 @@ export function createHttpOptionsConstructor() {
       this.headers = args.headers;
       this.httpRequestClient = args.httpRequestClient || fetchRequest;
       this.httpRequestInterceptors = args.httpRequestInterceptors;
-      this.enablePollDelay = !!args.enablePollDelay;
+      this.pollDelay = args.pollDelay;
     }
   };
 }