+++ test/spec/OktaAuth/constructor.ts
@@ -61,7 +61,7 @@ describe('OktaAuth (constructor)', () => {
     'pkce',
     'headers',
     'devMode',
-    'enablePollDelay',
+    'pollDelay',
     'ignoreSignature',
     'ignoreLifetime',
     'storageUtil',
@@ -70,7 +70,7 @@ describe('OktaAuth (constructor)', () => {
   it('saves expected options', () => {
     const config: any = {};
     savedOptions.forEach((option) => {
-      let val: string | object | boolean = 'fake_' + option; // default "fake" value
+      let val: string | object | boolean | number = 'fake_' + option; // default "fake" value
       switch (option) { // some types are strictly enforced. These should differ from the default
         case 'issuer':
         case 'redirectUri':
@@ -109,11 +109,13 @@ describe('OktaAuth (constructor)', () => {
         case 'ignoreSignature':
         case 'ignoreLifetime':
         case 'devMode':
-        case 'enablePollDelay':
           val = true;
           break;
         case 'pkce':
           val = false;
+          break;
+        case 'pollDelay':
+          val = 500;
           break;
         case 'scopes':
           val = [val];