+++ test/spec/authn/mfa-challenge.js
@@ -1552,7 +1552,7 @@ describe('MFA_CHALLENGE', function () {
     // OKTA-823470: iOS18 polling issue
     // NOTE: only run these tests in browser environments
     // eslint-disable-next-line no-extra-boolean-cast
-    (!!global.document ? describe : describe.skip)('enablePollDelay', () => {
+    (!!global.document ? describe : describe.skip)('pollDelay', () => {
       const togglePageVisibility = () => {
         document.hidden = !document.hidden;
         document.dispatchEvent(new Event('visibilitychange'));
@@ -1597,7 +1597,7 @@ describe('MFA_CHALLENGE', function () {
         const oktaAuth = new OktaAuth({
           issuer: 'https://auth-js-test.okta.com',
           httpRequestClient: context.httpSpy,
-          enablePollDelay: true
+          pollDelay: 500
         });
 
         context.transaction = oktaAuth.tx.createTransaction(mfaPush.response);